Auth dialog
added authentication when deleting or editing a record

+++ src/main/java/com/bitvault/ui/views/password/PasswordVM.java
@@ -3,10 +3,14 @@ package com.bitvault.ui.views.password;
 import com.bitvault.security.UserSession;
 import com.bitvault.services.interfaces.ICategoryService;
 import com.bitvault.services.interfaces.IPasswordService;
+import com.bitvault.ui.components.alert.ConfirmAlert;
+import com.bitvault.ui.components.alert.ErrorAlert;
+import com.bitvault.ui.exceptions.AuthenticateException;
 import com.bitvault.ui.model.Category;
 import com.bitvault.ui.model.Password;
 import com.bitvault.ui.model.Profile;
 import com.bitvault.ui.utils.JavaFxUtil;
+import com.bitvault.util.Labels;
 import com.bitvault.util.Result;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -17,14 +21,12 @@ import java.util.List;
 public class PasswordVM {
 
     private final UserSession userSession;
-    private final ObservableList<Password> passwords  = FXCollections.observableArrayList();
+    private final ObservableList<Password> passwords = FXCollections.observableArrayList();
     private final IPasswordService passwordService;
     private final ICategoryService categoryService;
     private final List<Category> categories;
 
-    public PasswordVM(
-            final UserSession userSession
-    ) {
+    public PasswordVM(final UserSession userSession) {
         this.userSession = userSession;
         this.passwordService = userSession.getServiceFactory().getPasswordService();
         this.categoryService = userSession.getServiceFactory().getCategoryService();
@@ -36,8 +38,7 @@ public class PasswordVM {
 
         Result<List<Password>> passwordsRes = passwordService.getPasswords();
         if (passwordsRes.isFail()) {
-            //TODO handle exception
-            passwordsRes.getError().printStackTrace();
+            ErrorAlert.show(Labels.i18n("error"), passwordsRes.getError());
             return;
         }
 
@@ -47,32 +48,34 @@ public class PasswordVM {
         Result<List<Category>> categoriesResult = categoryService.getCategories();
 
         if (categoriesResult.isFail()) {
-            //TODO handle exception
-            passwordsRes.getError().printStackTrace();
+            ErrorAlert.show(Labels.i18n("error"), passwordsRes.getError());
             return;
         }
         categories.clear();
         categories.addAll(categoriesResult.get());
-
-
     }
 
     public void create(Password password) {
         Result<Password> passwordResult = passwordService.create(password);
 
-        if(passwordResult.isFail()){
-            //TODO handle exception
+        if (passwordResult.isFail()) {
+            ErrorAlert.show(Labels.i18n("error"), passwordResult.getError());
             return;
         }
 
         passwords.add(passwordResult.get());
     }
 
+
     public void delete(Password password) {
-        passwordService.delete(password);
-        passwords.removeIf(
-                oldPass -> oldPass.getId().equals(password.getId())
-        );
+
+        boolean confirm = ConfirmAlert.deleteConfirm(userSession, password.getSecureDetails().isRequiresMp());
+
+        if (confirm) {
+            passwordService.delete(password);
+            passwords.removeIf(oldPass -> oldPass.getId().equals(password.getId()));
+        }
+
     }
 
     public void update(Password password) {
@@ -85,29 +88,36 @@ public class PasswordVM {
         });
     }
 
-    public boolean copyPassword(Password selectedItem){
+    public boolean copyPassword(Password selectedItem) {
         if (selectedItem == null) return false;
         final String decrypt = userSession.getEncryptionProvider().decrypt(selectedItem.getPassword());
         JavaFxUtil.copyToClipBoard(decrypt);
         return true;
     }
 
-    public boolean copyUsername(Password selectedItem){
+    public boolean copyUsername(Password selectedItem) {
         if (selectedItem == null) return false;
         JavaFxUtil.copyToClipBoard(selectedItem.getUsername());
         return true;
     }
 
-    public Password prepareForEdit(Password oldPass){
+    public Result<Password> prepareForEdit(Password oldPass) {
+        boolean confirm = ConfirmAlert.editAuthenticate(userSession, oldPass.getSecureDetails().isRequiresMp());
+
+        if (!confirm) {
+            return Result.error(AuthenticateException.INSTANCE);
+        }
+
         final String decrypt = userSession.getEncryptionProvider().decrypt(oldPass.getPassword());
 
-        return new Password(
+        final Password password = new Password(
                 oldPass.getId(),
                 oldPass.getUsername(),
                 decrypt,
                 oldPass.getSecureDetails()
         );
 
+        return Result.ok(password);
     }
 
     public ObservableList<Password> getPasswords() {
